+++ publish-cli/src/main/java/com/ericsson/eiffel/remrem/publish/cli/CliOptions.java
@@ -79,10 +79,12 @@ public class CliOptions {
         options.addOption("mp", "messaging_protocol", true, "name of messaging protocol to be used, e.g. eiffel3, eiffelsemantics, default is eiffelsemantics");
         options.addOption("domain", "domainId", true, "identifies the domain that produces the event");
         options.addOption("cc", "channelsCount", true, "Number of channels connected to message bus, default is 1");
+        options.addOption("wcto", "wait_for_confirms_timeOut", true, "the timeout for wait for confirms, default is 5000 ms/milliseconds");
         options.addOption("ud", "user_domain_suffix", true, "user domain suffix");
         options.addOption("v", "lists the versions of publish and all loaded protocols");
         options.addOption("tag", "tag", true, "tag to be used in routing key");
         options.addOption("rk", "routing_key", true, "routing key of the eiffel message. When provided routing key is not generated and the value provided is used.");
+        options.addOption("tto", "tcp_time_out", true, "specifies tcp connection timeout, default time is 60000 milliseconds");
         options.addOption("srkt", SEMANTICS_ROUTINGKEY_TYPE_OVERRIDE_FILEPATH, false, "routing key of the eiffel message. When provided routing key is not generated and the value provided is used.");
 
         contentGroup = createContentGroup();
@@ -220,10 +222,23 @@ public class CliOptions {
             String key = PropertiesConfig.DOMAIN_ID;
             System.setProperty(key, domain);
         }
+
         if (commandLine.hasOption("channelsCount")) {
             String channelsCount = commandLine.getOptionValue("channelsCount");
             String key = PropertiesConfig.CHANNELS_COUNT;
             System.setProperty(key, channelsCount);
+        }
+
+        if (commandLine.hasOption("tto")) {
+            String timeOut = commandLine.getOptionValue("tto");
+            String key = PropertiesConfig.TCP_TIMEOUT;
+            System.setProperty(key, timeOut);
+        }
+        
+        if (commandLine.hasOption("wcto")) {
+            String timeOut = commandLine.getOptionValue("wcto");
+            String key = PropertiesConfig.WAIT_FOR_CONFIRMS_TIME_OUT;
+            System.setProperty(key,timeOut);
         }
 
         if (commandLine.hasOption("tls")) {
@@ -279,6 +294,8 @@ public class CliOptions {
         System.clearProperty(PropertiesConfig.TLS);
         System.clearProperty(PropertiesConfig.DOMAIN_ID);
         System.clearProperty(PropertiesConfig.CHANNELS_COUNT);
+        System.clearProperty(PropertiesConfig.TCP_TIMEOUT);
+        System.clearProperty(PropertiesConfig.WAIT_FOR_CONFIRMS_TIME_OUT);
         System.clearProperty(PropertiesConfig.SEMANTICS_ROUTINGKEY_TYPE_OVERRIDE_FILEPATH);
     }