+++ publish-common/src/main/java/com/ericsson/eiffel/remrem/publish/config/PropertiesConfig.java
@@ -19,6 +19,8 @@ public class PropertiesConfig {
     public static final String MESSAGE_BUS_PORT = "com.ericsson.eiffel.remrem.publish.messagebus.port";
     public static final String VIRTUAL_HOST = "com.ericsson.eiffel.remrem.publish.messagebus.virtualhost";
     public static final String CHANNELS_COUNT = "com.ericsson.eiffel.remrem.publish.messagebus.channels";
+    public static final String TCP_TIMEOUT = "com.ericsson.eiffel.remrem.publish.messagebus.tcpTimeOut";
+    public static final String WAIT_FOR_CONFIRMS_TIME_OUT = "com.ericsson.eiffel.remrem.publish.messagebus.waitforconfirmstimeout";
     public static final String TLS = "com.ericsson.eiffel.remrem.publish.messagebus.tls";
     public static final String EXCHANGE_NAME = "com.ericsson.eiffel.remrem.publish.exchange.name";
     public static final String USE_PERSISTENCE = "com.ericsson.eiffel.remrem.publish.use.persistence";
@@ -46,7 +48,10 @@ public class PropertiesConfig {
     public static final String INVALID_EXCHANGE_MESSAGE_SERVICE = " ExchangeName is not present, To create the exchange specify createExchangeIfNotExisting in application configuration";
 
     public static final String SERVER_DOWN = "Internal Server Error";
+    public static final String GATEWAY_TIMEOUT = "Gateway Timeout";
     public static final String SERVER_DOWN_MESSAGE = "RabbitMQ is down. Please try later";
+    public static final String MESSAGE_NACK = "Message is nacked";
+    public static final String TIMEOUT_WAITING_FOR_ACK = "Time out waiting for ACK";
     public static final String ROUTING_KEY_GENERATION_FAILED_CONTENT = "Could not prepare Routing key to publish message";
     public static final String UNSUCCESSFUL_EVENT_CONTENT = "Please check previous event and try again later";
     public static final String RABBITMQ_PROPERTIES_NOT_FOUND = "RabbitMQ properties not found";